# [Opt](cast) Speed up new cast to datelike type