--- a/be/src/vec/functions/cast/cast_base.h
+++ b/be/src/vec/functions/cast/cast_base.h
@@ -19,6 +19,7 @@
 #include <cstddef>
 #include <cstdint>
 
+#include "cast_parameters.h"
 #include "vec/core/block.h"
 #include "vec/core/call_on_type_index.h"
 #include "vec/data_types/data_type.h"
@@ -41,6 +42,7 @@
 #include "vec/functions/function.h"
 #include "vec/functions/function_helpers.h"
 #include "vec/io/io_helper.h"
+
 namespace doris::vectorized {
 
 struct NameCast {
@@ -166,11 +168,6 @@ public:
                 cast_mode_type_to_string(CastMode, block.get_by_position(arguments[0]).type,
                                          block.get_by_position(result).type));
     }
-};
-
-struct CastParameters {
-    Status status = Status::OK();
-    bool is_strict = false;
 };
 
 #ifdef BE_TEST