… be/src/vec/functions/cast/cast_to_date_or_datetime_impl.hpp
… be/src/vec/functions/cast/cast_to_datetimev2_impl.hpp
… be/src/vec/functions/cast/cast_to_datev2_impl.hpp
… be/test/vec/function/cast/cast_to_time_test.cpp
… regression-test/data/cast_p0/cast_to_time.out