### What problem does this PR solve?

Issue Number: close #xxx

Related PR: #xxx

Problem Summary:

based on https://github.com/apache/doris/pull/54357. wait it merged firstly.
compared with 3.0.6:

master
```sql
mysql> SELECT COUNT(CAST(date_iso AS DATE)) AS d
    -> FROM test_string_datetime;
+-----------+
| d         |
+-----------+
| 800000000 |
+-----------+
1 row in set (21.04 sec)
mysql> SELECT COUNT(*) FROM (
    ->   SELECT CAST(datetime_iso AS DATETIME) AS dt
    ->   FROM test_string_datetime
    -> ) t;
+-----------+
| COUNT(*)  |
+-----------+
| 800000000 |
+-----------+
1 row in set (0.29 sec)
mysql> SELECT COUNT(*) FROM (
    ->   SELECT CAST(datetime_offset AS DATETIME) AS dt
    ->   FROM test_string_datetime
    -> ) t;
+-----------+
| COUNT(*)  |
+-----------+
| 800000000 |
+-----------+
1 row in set (0.29 sec)
mysql> SELECT COUNT(*) FROM (
    ->   SELECT CAST(datetime_named_tz AS DATETIME) AS dt
    ->   FROM test_string_datetime
    -> ) t;
+-----------+
| COUNT(*)  |
+-----------+
| 800000000 |
+-----------+
1 row in set (0.29 sec)
```


3.0.6
```sql
mysql> SELECT COUNT(CAST(date_iso AS DATE)) AS d FROM test_string_datetime;
+-----------+
| d         |
+-----------+
| 800000000 |
+-----------+
1 row in set (28.29 sec)
mysql> SELECT COUNT(*) FROM (
    ->   SELECT CAST(datetime_iso AS DATETIME) AS dt
    ->   FROM test_string_datetime
    -> ) t;
+-----------+
| COUNT(*)  |
+-----------+
| 800000000 |
+-----------+
1 row in set (0.21 sec)
mysql> SELECT COUNT(*) FROM (
    ->   SELECT CAST(datetime_offset AS DATETIME) AS dt
    ->   FROM test_string_datetime
    -> ) t;
+-----------+
| COUNT(*)  |
+-----------+
| 800000000 |
+-----------+
1 row in set (0.22 sec)
mysql> SELECT COUNT(*) FROM (
    ->   SELECT CAST(datetime_named_tz AS DATETIME) AS dt
    ->   FROM test_string_datetime
    -> ) t;
+-----------+
| COUNT(*)  |
+-----------+
| 800000000 |
+-----------+
1 row in set (0.22 sec)
```

### Release note

None

### Check List (For Author)

- Test <!-- At least one of them must be included. -->
    - [ ] Regression test
    - [ ] Unit Test
    - [ ] Manual test (add detailed scripts or steps below)
    - [x] No need to test or manual test. Explain why:
        - [x] This is a refactor/code format and no logic has been changed.
        - [ ] Previous test can cover this change.
        - [ ] No code files have been changed.
        - [ ] Other reason <!-- Add your reason?  -->

- Behavior changed:
    - [x] No.
    - [ ] Yes. <!-- Explain the behavior change -->

- Does this need documentation?
    - [x] No.
    - [ ] Yes. <!-- Add document PR link here. eg: https://github.com/apache/doris-website/pull/1214 -->

### Check List (For Reviewer who merge this PR)

- [ ] Confirm the release note
- [ ] Confirm test cases
- [ ] Confirm document
- [ ] Add branch pick label <!-- Add branch pick label that this PR should merge into -->