@hello-stephen 2025-08-09
# BE Regression && UT Coverage Report
Increment line coverage `90.93% (361/397)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/54359_b24a507a4117b80cd04a3b3568051d2845200c5c_merge/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/54359_b24a507a4117b80cd04a3b3568051d2845200c5c_merge/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage | 81.35% (22492/27650) |
| Line Coverage     | 74.06% (233141/314781) |
| Region Coverage   | 61.44% (193889/315579) |
| Branch Coverage   | 65.55% (83770/127791) |